requirements.txt
streamlit>=1.45.1
streamlit-webrtc>=0.63.1
pandas
pydub>=0.25.1
google-genai>=1.19.0
pydantic
sqlalchemy>=2.0.41
cloud-sql-python-connector[pymysql]>=1.18.2
python-dotenv>=1.1.0
google-cloud-documentai
numpy
scipy
requests
google-cloud-documentai>=3.5.0
Authlib>=1.3.2
audioop-lts